+++ astropy_healpix/interpolation.h
@@ -1,0 +1,12 @@
+#ifdef _MSC_VER
+#if _MSC_VER >= 1600
+#include <stdint.h>
+#else
+#include <stdint_msc.h>
+#endif
+#else
+#include <stdint.h>
+#endif
+
+void interpolate_weights(double lon, double lat, int64_t *ring_indices,
+                         double *weights, int Nside);