+++ astropy_healpix/tests/test_core.py
@@ -17,7 +17,8 @@ from ..core import (nside_to_pixel_area, nside_to_pixel_resolution,
                     nside_to_npix, npix_to_nside, healpix_to_lonlat,
                     lonlat_to_healpix, interpolate_bilinear_lonlat,
                     neighbours, healpix_cone_search, boundaries_lonlat,
-                    level_to_nside, nested_to_ring, ring_to_nested)
+                    level_to_nside, nested_to_ring, ring_to_nested,
+                    bilinear_interpolation_weights)
 
 
 def test_level_to_nside():
@@ -163,6 +164,40 @@ def test_nested_ring_shape(function):
 
 
 @pytest.mark.parametrize('order', ['nested', 'ring'])
+def test_bilinear_interpolation_weights(order):
+
+    indices, weights = bilinear_interpolation_weights(100 * u.deg, 10 * u.deg,
+                                                      nside=4, order=order)
+    if order == 'nested':
+        indices = nested_to_ring(indices, nside=4)
+    assert_equal(indices, [76, 77, 60, 59])
+    assert_allclose(weights, [0.532723, 0.426179, 0.038815, 0.002283], atol=1e-6)
+
+
+def test_bilinear_interpolation_weights_invalid():
+    with pytest.raises(ValueError) as exc:
+        bilinear_interpolation_weights(1 * u.deg, 2 * u.deg, nside=5)
+    assert exc.value.args[0] == 'nside must be a power of two'
+
+    with pytest.raises(ValueError) as exc:
+        bilinear_interpolation_weights(3 * u.deg, 4 * u.deg,
+                                       nside=4, order='banana')
+    assert exc.value.args[0] == "order must be 'nested' or 'ring'"
+
+
+def test_bilinear_interpolation_weights_shape():
+
+    indices, weights = bilinear_interpolation_weights(3 * u.deg, 4 * u.deg, nside=8)
+    assert indices.shape == (4,)
+    assert weights.shape == (4,)
+
+    indices, weights = bilinear_interpolation_weights([[1, 2, 3], [2, 3, 4]] * u.deg,
+                                                      [[1, 2, 3], [2, 3, 4]] * u.deg, nside=8)
+    assert indices.shape == (4, 2, 3)
+    assert weights.shape == (4, 2, 3)
+
+
+@pytest.mark.parametrize('order', ['nested', 'ring'])
 def test_interpolate_bilinear_lonlat(order):
     values = np.ones(192) * 3
     result = interpolate_bilinear_lonlat([1, 3, 4] * u.deg, [3, 2, 6] * u.deg,
@@ -184,6 +219,7 @@ def test_interpolate_bilinear_invalid():
 
 
 def test_interpolate_bilinear_lonlat_shape():
+
     values = np.ones(192) * 3
 
     result = interpolate_bilinear_lonlat(3 * u.deg, 4 * u.deg, values)
@@ -192,6 +228,14 @@ def test_interpolate_bilinear_lonlat_shape():
     result = interpolate_bilinear_lonlat([[1, 2, 3], [2, 3, 4]] * u.deg,
                                          [[1, 2, 3], [2, 3, 4]] * u.deg, values)
     assert result.shape == (2, 3)
+
+    values = np.ones((192, 50)) * 3
+
+    lon = np.ones((3, 6, 5)) * u.deg
+    lat = np.ones((3, 6, 5)) * u.deg
+
+    result = interpolate_bilinear_lonlat(lon, lat, values)
+    assert result.shape == (3, 6, 5, 50)
 
 
 @pytest.mark.parametrize('order', ['nested', 'ring'])